+++ packages/tgweb/lib/tgweb/generate_html.mjs
@@ -2,7 +2,7 @@ import * as PATH from "path"
 import getType from "./get_type.mjs"
 import generationFuncs from "./generation_funcs.mjs"
 
-const generateHTML = function(path, siteData) {
+const generateHTML = (path, siteData) => {
   const type = getType(PATH.dirname(path))
 
   if (typeof generationFuncs[type] === "function") return generationFuncs[type](path, siteData)