+++ packages/tgweb/lib/tgweb/generation_funcs.mjs
@@ -2,145 +2,147 @@ import * as PATH from "path"
 import pretty from "pretty"
 import { JSDOM } from "jsdom"
 import { minimatch } from "minimatch"
+import { setTgAttrs } from "./set_tg_attrs.mjs"
+import { removeTgAttributes } from "./remove_tg_attributes.mjs"
 
 const generationFuncs = {}
 
-generationFuncs["page"] = function(path, siteData) {
+generationFuncs["page"] = (path, siteData) => {
   const filename = PATH.basename(path)
   const page = siteData.pages.find(page => page.path == filename)
 
   if (page) {
     const pageRoot = page.dom.window.document.body.children[0].cloneNode(true)
-    const tgLayoutAttr = pageRoot.attributes.getNamedItem("tg-layout")
+    setTgAttrs(pageRoot)
     const headAttrs = { title: getTitle(pageRoot) }
 
-    removeSpecialAttributes(pageRoot)
     embedComponents(pageRoot, siteData)
     embedArticles(pageRoot, siteData)
     embedArticleLists(pageRoot, siteData)
     embedLinksToArticles(pageRoot, siteData, path)
 
-    if (tgLayoutAttr) {
-      const layoutName = tgLayoutAttr.value
-
-      const layout = siteData.layouts.find(layout => layout.path == layoutName + ".html")
-
-      if (layout) {
-        const layoutRoot = layout.dom.window.document.body.cloneNode(true)
-        embedComponents(layoutRoot, siteData)
-
-        const target = layoutRoot.querySelector("[tg-content]")
-
-        if (target) target.replaceWith(pageRoot)
-
-        return renderHTML(layoutRoot, siteData, headAttrs)
-      }
-    }
-    else {
-      return renderHTML(pageRoot, siteData, headAttrs)
-    }
+    const layoutRoot = applyLayout(pageRoot, siteData)
+    if (layoutRoot) return renderHTML(layoutRoot, siteData, headAttrs)
+    else return renderHTML(pageRoot, siteData, headAttrs)
   }
 
   return pretty(siteData.documentTemplate.serialize())
 }
 
-generationFuncs["article"] = function(path, siteData) {
+generationFuncs["article"] = (path, siteData) => {
   const article = siteData.articles.find(article => "src/articles/" + article.path == path)
 
   if (article) {
     const articleRoot = article.dom.window.document.body.children[0].cloneNode(true)
-    const tgLayoutAttr = articleRoot.attributes.getNamedItem("tg-layout")
-    const headAttrs = { title: getTitle(articleRoot) }
+    setTgAttrs(articleRoot)
 
-    removeSpecialAttributes(articleRoot)
-    embedComponents(articleRoot, siteData)
-    embedLinksToArticles(articleRoot, siteData, path)
-
-    if (tgLayoutAttr) {
-      const layoutName = tgLayoutAttr.value
-
-      const layout = siteData.layouts.find(layout => layout.path == layoutName + ".html")
-
-      if (layout) {
-        const layoutRoot = layout.dom.window.document.body.cloneNode(true)
-        embedComponents(layoutRoot, siteData)
-
-        const target = layoutRoot.querySelector("[tg-content]")
-
-        if (target) target.replaceWith(articleRoot)
-
-        return renderHTML(layoutRoot, siteData, headAttrs)
-      }
+    if (articleRoot.tgAttrs["component"]) {
+      const componentRoot = getComponentRoot(articleRoot, siteData)
+      setTgAttrs(componentRoot)
+      return renderArticle(componentRoot, siteData, path)
     }
     else {
-      return renderHTML(articleRoot, siteData, headAttrs)
+      embedComponents(articleRoot, siteData)
+      return renderArticle(articleRoot, siteData, path)
     }
   }
 
   return pretty(siteData.documentTemplate.serialize())
 }
 
-const embedComponents = function(node, siteData) {
+const renderArticle = (articleRoot, siteData, path) => {
+  const headAttrs = { title: getTitle(articleRoot) }
+  embedLinksToArticles(articleRoot, siteData, path)
+  const layoutRoot = applyLayout(articleRoot, siteData)
+  if (layoutRoot) return renderHTML(layoutRoot, siteData, headAttrs)
+}
+
+const applyLayout = (element, siteData) => {
+  if (element.tgAttrs["layout"] === undefined) return
+
+  const layout =
+    siteData.layouts.find(layout => layout.path == element.tgAttrs["layout"] + ".html")
+
+  if (layout === undefined) return
+
+  const layoutRoot = layout.dom.window.document.body.cloneNode(true)
+  embedComponents(layoutRoot, siteData)
+
+  const target = layoutRoot.querySelector("[tg-content]")
+
+  if (target) target.replaceWith(element)
+
+  return layoutRoot
+}
+
+const embedComponents = (node, siteData) => {
   const targets = node.querySelectorAll("[tg-component]")
 
   targets.forEach(target => {
-    const tgComponentAttr = target.attributes.getNamedItem("tg-component")
-    const componentName = tgComponentAttr.value
-    const args = {}
-
-    for (let i = 0; i < target.attributes.length; i++) {
-      const attr = target.attributes.item(i)
-      const parts = attr.name.split("-")
-
-      if (parts.length == 3 && parts[0] == "tg" && parts[1] == "data") {
-        args[parts[2]] = attr.value
-      }
-    }
-
-    const component =
-      siteData.components.find(component => component.path == componentName + ".html")
-
-    if (component) {
-      const componentRoot = component.dom.window.document.body.children[0].cloneNode(true)
-      embedArgs(componentRoot, args)
-      target.replaceWith(componentRoot)
-    }
+    setTgAttrs(target)
+    const componentRoot = getComponentRoot(target, siteData)
+    if (componentRoot) target.replaceWith(componentRoot)
   })
 }
 
-const embedArticles = function(node, siteData) {
+const getComponentRoot = (element, siteData) => {
+  const componentName = element.tgAttrs["component"]
+
+  const component =
+    siteData.components.find(component => component.path == componentName + ".html")
+
+  if (component) {
+    const componentRoot = component.dom.window.document.body.children[0].cloneNode(true)
+    const args = getArgs(element)
+    embedArgs(componentRoot, args)
+    return componentRoot
+  }
+}
+
+const getArgs = element => {
+  const args = {}
+  const names = ["title", "description", "index", "date"]
+
+  names.forEach(name => {
+    if (element.tgAttrs[name]) args[name] = element.tgAttrs[name]
+  })
+
+  for (const key in element.tgAttrs.data) args[key] = element.tgAttrs.data[key]
+
+  return args
+}
+
+const embedArticles = (node, siteData) => {
   const targets = node.querySelectorAll("[tg-article]")
 
   targets.forEach(target => {
-    const tgArticleAttr = target.attributes.getNamedItem("tg-article")
-    const articleName = tgArticleAttr.value
+    setTgAttrs(target)
 
     const article =
-      siteData.articles.find(article => article.path == articleName + ".html")
+      siteData.articles.find(article => article.path == target.tgAttrs["article"] + ".html")
 
     if (article) {
       const articleRoot = article.dom.window.document.body.children[0].cloneNode(true)
-      removeSpecialAttributes(articleRoot)
+      setTgAttrs(articleRoot)
       target.replaceWith(articleRoot)
     }
   })
 }
 
-const embedArticleLists = function(node, siteData) {
+const embedArticleLists = (node, siteData) => {
   const targets = node.querySelectorAll("[tg-articles]")
 
   targets.forEach(target => {
-    const pattern = target.attributes.getNamedItem("tg-articles").value
+    setTgAttrs(target)
+    const pattern = target.tgAttrs["articles"]
     const tag = getTag(target)
-    const tgOrderByAttr = target.attributes.getNamedItem("tg-order-by")
-
     const articles = filterArticles(siteData.articles, pattern, tag)
 
-    if (tgOrderByAttr) sortArticles(articles, tgOrderByAttr.value)
+    if (target.tgAttrs["order-by"]) sortArticles(articles, target.tgAttrs["order-by"])
 
     articles.forEach(article => {
       const articleRoot = article.dom.window.document.body.children[0].cloneNode(true)
-      removeSpecialAttributes(articleRoot)
+      setTgAttrs(articleRoot)
       target.before(articleRoot)
     })
 
@@ -148,20 +150,21 @@ const embedArticleLists = function(node, siteData) {
   })
 }
 
-const embedLinksToArticles = function(node, siteData, path) {
+const embedLinksToArticles = (node, siteData, path) => {
   const targets = node.querySelectorAll("[tg-links]")
 
   targets.forEach(target => {
-    const pattern = target.attributes.getNamedItem("tg-links").value
+    setTgAttrs(target)
+    const pattern = target.tgAttrs["links"]
     const tag = getTag(target)
-    const tgOrderByAttr = target.attributes.getNamedItem("tg-order-by")
 
     const articles = filterArticles(siteData.articles, pattern, tag)
 
-    if (tgOrderByAttr) sortArticles(articles, tgOrderByAttr.value)
+    if (target.tgAttrs["order-by"]) sortArticles(articles, target.tgAttrs["order-by"])
 
     articles.forEach(article => {
       const articleRoot = article.dom.window.document.body.children[0]
+      setTgAttrs(articleRoot)
       const copy = target.cloneNode(true)
 
       const href = PATH.relative(PATH.dirname(path), PATH.join("src/articles", article.path))
@@ -169,12 +172,10 @@ const embedLinksToArticles = function(node, siteData, path) {
 
       const args = {
         "title": getTitle(articleRoot),
-        "date": getDate(articleRoot)
+        "date": articleRoot.tgAttrs["date"]
       }
 
       embedArgs(copy, args)
-
-      removeSpecialAttributes(copy)
 
       target.before(copy)
     })
@@ -183,16 +184,16 @@ const embedLinksToArticles = function(node, siteData, path) {
   })
 }
 
-const filterArticles = function(articles, pattern, tag) {
+const filterArticles = (articles, pattern, tag) => {
   articles =
     articles.filter(article => {
       if (minimatch(article.path, pattern)) {
         if (tag) {
           const articleRoot = article.dom.window.document.body.children[0]
-          const tgTagAttr = articleRoot.attributes.getNamedItem("tg-tag")
+          setTgAttrs(articleRoot)
 
-          if (tgTagAttr) {
-            return tgTagAttr.value == tag
+          if (articleRoot.tgAttrs["tags"]) {
+            return articleRoot.tgAttrs["tags"].split(",").includes(tag)
           }
         }
         else {
@@ -204,7 +205,7 @@ const filterArticles = function(articles, pattern, tag) {
   return articles
 }
 
-const sortArticles = function(articles, orderBy) {
+const sortArticles = (articles, orderBy) => {
   const re = /^(index):(asc|desc)$/
   const md = re.exec(orderBy)
 
@@ -214,15 +215,17 @@ const sortArticles = function(articles, orderBy) {
       const d = b.dom.window.document.body.children[0]
 
       if (c && d) {
-        const i = c.attributes.getNamedItem("tg-index")
-        const j = d.attributes.getNamedItem("tg-index")
+        setTgAttrs(c)
+        setTgAttrs(d)
+        const i = c.tgAttrs["index"]
+        const j = d.tgAttrs["index"]
 
         if (i) {
           if (j) {
-            if (i.value > j.value) return -1
-            if (i.value < j.value) return 1
-            if (i.path > j.path) return -1
-            if (i.path < j.path) return 1
+            if (i > j) return -1
+            if (i < j) return 1
+            if (a.path > b.path) return -1
+            if (a.path < b.path) return 1
             return 0
           }
           else {
@@ -244,12 +247,12 @@ const sortArticles = function(articles, orderBy) {
   }
 }
 
-const embedArgs = function(node, args) {
+const embedArgs = (node, args) => {
   const targets = node.querySelectorAll("[tg-text]")
 
   targets.forEach(target => {
-    const tgTextAttr = target.attributes.getNamedItem("tg-text")
-    const text = tgTextAttr.value
+    setTgAttrs(target)
+    const text = target.tgAttrs["text"]
 
     if (text.substring(0, 2) == "%{" && text.substring(text.length - 1) == "}") {
       const paramName = text.substring(2, text.length - 1)
@@ -258,13 +261,12 @@ const embedArgs = function(node, args) {
         target.innerHTML = args[paramName]
       }
     }
-
-    removeSpecialAttributes(target)
   })
 }
 
-const renderHTML = function(root, siteData, headAttrs) {
+const renderHTML = (root, siteData, headAttrs) => {
   const dom = new JSDOM(siteData.documentTemplate.serialize())
+  removeTgAttributes(root)
   dom.window.document.body.replaceWith(root)
 
   if (headAttrs["title"])
@@ -273,20 +275,8 @@ const renderHTML = function(root, siteData, headAttrs) {
   return pretty(dom.serialize())
 }
 
-const removeSpecialAttributes = function(element) {
-  const names = []
-
-  for (let i = 0; i < element.attributes.length; i++) {
-    const attr = element.attributes[i]
-    if (attr.name.match(/^tg-/)) names.push(attr.name)
-  }
-
-  names.forEach(n => element.attributes.removeNamedItem(n))
-}
-
-const getTitle = function(element) {
-  const title = element.attributes.getNamedItem("tg-title")
-  if (title) return title.value
+const getTitle = element => {
+  if (element.tgAttrs["title"]) return element.tgAttrs["title"]
 
   const h1 = element.querySelector("h1")
   if (h1) return h1.textContent
@@ -307,17 +297,12 @@ const getTitle = function(element) {
   if (h6) return h6.textContent
 }
 
-const getDate = function(element) {
-  const dateAttr = element.attributes.getNamedItem("tg-date")
-  if (dateAttr) return dateAttr.value
-}
+const getTag = element => {
+  element.tgAttrs["filter"]
 
-const getTag = function(element) {
-  const tgFilterAttr = element.attributes.getNamedItem("tg-filter")
-
-  if (tgFilterAttr) {
+  if (element.tgAttrs["filter"]) {
     const re = /^(tag):(.+)$/
-    const md = re.exec(tgFilterAttr.value)
+    const md = re.exec(element.tgAttrs["filter"])
     if (md) return md[2]
   }
 }