+++ packages/tgweb/CHANGELOG.md
@@ -1,5 +1,10 @@
 # tgweb CHANGELOG
 
+## 0.0.4
+
+* d4f1cd4 Use `tg-tags` instead of `tg-tag` attribute
+* ca1bd73 Merge attributes of article and component
+
 ## 0.0.3
 
-- b4769b2 Generate a list of links to articles with a specific tag
+* b4769b2 Generate a list of links to articles with a specific tag